+++ empower-italy/components/Footer.tsx
@@ -1,26 +1,35 @@
+'use client'
+
 import Link from "next/link";
+import { useSelectedLayoutSegment } from "next/navigation";
 
 export default function Footer() {
+
+    const segment = useSelectedLayoutSegment();
+
+
+                            // <Link key={link.name} className={`/${segment}` === link.href ? 'nav-button-active' : 'nav-button'} href={link.href}></Link>
+
     return (
         <section id="footer" className="pb-6">
             <article className="max-w-7xl m-auto">
                 <div className="grid grid-cols-12 gap-6 h-full">
                         
                     <div className="group col-span-4 border-2 border-indigo-600 rounded-2xl p-12 hover:bg-indigo-600 transition-all">
-                        <Link href="/workflow" className="font-bold text-4xl text-indigo-600 uppercase group-hover:text-gray-50 transition-all">
-                            Go to workflow
+                        <Link href={`${segment}` === 'workflow' ? '/' : '/workflow'} className="font-bold text-4xl text-indigo-600 uppercase group-hover:text-gray-50 transition-all">
+                            Go to {`${segment}` === 'workflow' ? 'home' : 'workflow'}
                         </Link>
                     </div>
                         
                     <div className="group col-span-4 border-2 border-indigo-600 rounded-2xl p-12 hover:bg-indigo-600 transition-all">
-                        <Link href="/documentation" className="font-bold text-4xl text-indigo-600 uppercase group-hover:text-gray-50 transition-all">
-                            Go to documentation
+                        <Link href={`${segment}` === 'documentation' ? '/' : '/documentation'} className="font-bold text-4xl text-indigo-600 uppercase group-hover:text-gray-50 transition-all">
+                            Go to {`${segment}` === 'documentation' ? 'home' : 'documentation'}
                         </Link>
                     </div>
                         
                     <div className="group col-span-4 border-2 border-indigo-600 rounded-2xl p-12 hover:bg-indigo-600 transition-all">
-                        <Link href="/datasets" className="font-bold text-4xl text-indigo-600 uppercase group-hover:text-gray-50 transition-all">
-                            Go to datasets
+                        <Link href={`${segment}` === 'datasets' ? '/' : '/datasets'} className="font-bold text-4xl text-indigo-600 uppercase group-hover:text-gray-50 transition-all">
+                            Go to {`${segment}` === 'datasets' ? 'home' : 'datasets'}
                         </Link>            
                     </div>
                     <div className="col-span-12 pt-16 pb-10 text-center font-bold text-xl text-indigo-600">